feat: move tenants service to proxy entry point
