+++ npm/ng-packs/packages/tenant-management/proxy/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../../dist/packages/tenant-management/proxy",
+  "lib": {
+    "entryFile": "src/public-api.ts"
+  }
+}
