+++ npm/ng-packs/packages/tenant-management/proxy/src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './proxy';
+
