+++ npm/ng-packs/packages/tenant-management/proxy/src/public-api.ts
@@ -1,0 +1,1 @@
+export * from './lib';
