+++ npm/ng-packs/packages/tenant-management/src/lib/defaults/default-tenants-entity-props.ts
@@ -1,5 +1,5 @@
+import { TenantDto } from '@abp/ng.tenant-management/proxy';
 import { EntityProp, ePropType } from '@abp/ng.theme.shared/extensions';
-import { TenantDto } from '../proxy/models';
 
 export const DEFAULT_TENANTS_ENTITY_PROPS = EntityProp.createMany<TenantDto>([
   {
