+++ npm/ng-packs/packages/tenant-management/src/lib/models/tenant-management.ts
@@ -1,5 +1,5 @@
 import { PagedResultDto } from '@abp/ng.core';
-import { TenantDto } from '../proxy/models';
+import { TenantDto } from '@abp/ng.tenant-management/proxy';
 
 export namespace TenantManagement {
   export interface State {
