+++ npm/ng-packs/packages/tenant-management/src/public-api.ts
@@ -2,6 +2,5 @@ export * from './lib/components';
 export * from './lib/enums';
 export * from './lib/guards';
 export * from './lib/models';
-export * from './lib/proxy';
 export * from './lib/tenant-management.module';
 export * from './lib/tokens';
